requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
pypdf
faiss-cpu